fix: implementing suggestions
